Exercise5

+++ common/Adc.hpp
@@ -1,8 +1,8 @@
 #ifndef MICROTECH_ADC_HPP
 #define MICROTECH_ADC_HPP
 
+#include "MovingAverage.hpp"
 #include "helpers.hpp"
-
 #include <msp430g2553.h>
 #include <cstdint>
 #include <map>
@@ -87,7 +87,7 @@ public:
     // without the user having to actively fetch any data from the ADC10MEM.
     ADC10DTC0 = ADC10CT;                                   // enable continuous transfer
     ADC10DTC1 = sizeof(adcValues) / sizeof(adcValues[0]);  // Number of transfers is equal to the size of array.
-    ADC10SA = (uint16_t) & (adcValues[0]);                 // Starts at address is the first entry of the array;
+    ADC10SA = (size_t)(&adcValues[0]);                     // Starts at address is the first entry of the array;
   }
 
   /**
